Move JobStatusWriter to own module. Make collapsible if running in github actions.
