+++ src/rkojob/factories.py
@@ -1,4 +1,3 @@
-import os
 import sys
 
 from rkojob import JobContext, JobRunner
@@ -7,10 +6,12 @@ from rkojob import JobContext, JobRunner
 class JobContextFactory:
     @classmethod
     def create(cls, *args, **kwargs) -> JobContext:
-        from rkojob.context import JobContextImpl, JobStatusWriter
+        from rkojob.context import JobContextImpl
+        from rkojob.writer import JobStatusWriter
 
         status_writer: JobStatusWriter = kwargs.get("status_writer") or JobStatusWriter(
-            stream=sys.stdout, show_detail=False, collapsible_output=bool(os.getenv("GITHUB_ACTIONS"))
+            stream=sys.stdout,
+            show_detail=False,
         )
         return JobContextImpl(values=kwargs.get("values"), status_writer=status_writer)
